Fix previewer for PowerShell
PowerShell scripts (ps1 files) are not executable by default. This means setting a .ps1 file as the previewer will not work. This led to errors like: fork/exec C:\Users\USERNAME\Invoke-LfPreview.ps1: %1 is not a valid Win32 application.

This change makes it so the shell and shellflag are the first and second parameters that are passed in before the previewer script.