Check Go version is < 1.19 in `go generate`
This commit prevents `go generate` from running on Go
1.19. Go versions between Go 1.14 (where the `go generate`
command was introduced) and 1.18 are allowed.

Starting with Go 1.19, the `go doc` command changed the
format of its output. The `gen/man.sh` script fails to
parse the new output and generates a badly formatted
manpage. The `gen/docstring.sh` script generates a
readable docstring, but it's different from what
it was before.

This is a temporary and imperfect fix. A better fix
would be nice, but it might be easiest to wait until
we can require Go 1.19 for building `lf`. Otherwise,
we would need a solution that works the same
on both Go 1.18 and Go 1.19.

Fixes https://github.com/gokcehan/lf/issues/959.