For directories in preview column, dim cursor (#924)
* For directories in preview column, dim cursor

When preview is on and consists of a directory listing, make the
cursor in there dimmer to emphasize it's not active. This makes it
visually obvious which cursor is active at all times, without
having to read the path. The active cursor is now always  the
rightmost bright one, regardless of whether the `preview` option
is on or off.

See screenshots in the pull request at
https://github.com/gokcehan/lf/pull/924.

Before this, I noticed that I often open a file by accident because
I got confused which of the three cursors on the screen is the active
one.

* No-op changes to ui.go

Reduce number of arguments to printDir and list non-configurable colors