shell-pipe: fix excessive waiting for output
There are some programs that keep their stdin/stdout/stderr open beyond
their process lifetime. An example is "wl-copy", which forks off into
the background and keeps stderr open.

Before this change, the behaviour of lf's shell-pipe (aka %) was to
first wait for an EOF from the stdout/stderr pipe of the process, and
only then to wait for the process to exit.

As a result, a command such as "%wl-copy foo" never stops showing the
">" command prefix indicator.

Fix this, by waiting for the EOF from the pipe and the process exit in
parallel, as well as setting the "exec.Cmd.WaitDelay" field.

(For some reason, the fix works even without setting WaitDelay, and it
also doesn't seem to wait for one second; I haven't investigated why
this is; the behaviour is an improvement either way.)

Additionally, improve handling errors when reading from the
stdout/stderr pipe:
- Always give up reading from the pipe, regardless of error kind
- Unless it is an expected error (EOF or (now) ErrClosed), log the error
- Compare errors using "errors.Is"