.github: staticcheck-action → golangci-lint-action
golangci-lint bundles many Go linters
(https://golangci-lint.run/docs/linters/). For now, we are going with its
default set of linters, which includes staticcheck, but with the
errcheck linter explicitly excluded because it generates a lot of lints;
I plan to address those and add errcheck in a subsequent contribution.